fix: address all PR #45 review findings
🔴 Must Fix:
- Extract collaboration bridge from preload.ts into preload/collaboration.ts
  (removes inline import() type expressions violating project conventions)
- Extract models + google/imagegen bridges to keep preload.ts at 497 LOC
- Split electron.d.ts: workspace tool interfaces → electron-workspace.d.ts (431 LOC)
- Remove dead clientMessageId parameter from collaboration IPC handler,
  preload bridge, store, and type declarations
- Deduplicate types: electron/collaboration/ now imports CollaborationRole
  and CollaborationResult from shared src/types/collaboration.ts

🟡 Should Fix:
- Add durationMs to collab_specialist_end event type and forward real
  duration from engine → IPC → store → UI (was hardcoded to 0)
- Handle collab_specialist_start in switch + add exhaustive default: never
  check for compile-time safety on future event types
- Fix misleading status banner: show 'Research ✓ — Analyst & Visionary
  analyzing (N/2)' instead of confusing N/3 count
- Extract shared TERMINAL_STATUSES constant from OrchestrationPanel.tsx
  and subagent.ts into subagent-constants.ts

🟢 Nits:
- Soften Tavily requirement in researcher.md ('may be available' vs MUST)
- Restore blank lines between agent store methods for readability
