+++ apps/desktop/electron/preload.ts
@@ -5,6 +5,9 @@ import { debugBridge, lspBridge } from './preload/debug-lsp';
 import { subagentBridge } from './preload/subagent';
 import { skillsBridge } from './preload/skills';
 import { promptsBridge } from './preload/prompts';
+import { collaborationBridge } from './preload/collaboration';
+import { modelsBridge } from './preload/models';
+import { googleBridge, imagegenBridge } from './preload/google-imagegen';
 import type {
   WorkspaceInfo,
   WorkspaceConfig,
@@ -217,31 +220,10 @@ contextBridge.exposeInMainWorld('sero', {
       ipcRenderer.invoke(IpcChannels.appAgent.prompt, appId, workspaceId, text),
   },
 
-  models: {
-    list: (): Promise<import('../src/types/ipc').AvailableModelGroup[]> =>
-      ipcRenderer.invoke(IpcChannels.models.list),
-  },
+  models: modelsBridge,
 
-  google: {
-    execute: (service: string, subArgs: string[]) => ipcRenderer.invoke(IpcChannels.google.execute, service, subArgs),
-    authStatus: () => ipcRenderer.invoke(IpcChannels.google.authStatus),
-    login: () => ipcRenderer.invoke(IpcChannels.google.login),
-    logout: () => ipcRenderer.invoke(IpcChannels.google.logout),
-    onAuthEvent: (cb: (event: any) => void) => {
-      const handler = (_e: IpcRendererEvent, event: any) => cb(event);
-      ipcRenderer.on(IpcChannels.google.authEvent, handler);
-      return () => { ipcRenderer.removeListener(IpcChannels.google.authEvent, handler); };
-    },
-  },
-
-  imagegen: {
-    generate: (workspaceId: string, params: any): Promise<any> =>
-      ipcRenderer.invoke(IpcChannels.imagegen.generate, workspaceId, params),
-    readImage: (filePath: string): Promise<string> =>
-      ipcRenderer.invoke(IpcChannels.imagegen.readImage, filePath),
-    deleteImage: (workspaceId: string, generationId: number, singleImageId?: string): Promise<{ ok: boolean; error?: string }> =>
-      ipcRenderer.invoke(IpcChannels.imagegen.deleteImage, workspaceId, generationId, singleImageId),
-  },
+  google: googleBridge,
+  imagegen: imagegenBridge,
 
   voice: {
     status: (): Promise<VoiceTranscriptionStatus> =>
@@ -469,20 +451,7 @@ contextBridge.exposeInMainWorld('sero', {
       ipcRenderer.invoke(IpcChannels.feedback.remove, messageId),
   },
 
-  collaboration: {
-    prompt: (sessionId: string, workspaceId: string, query: string, clientMessageId?: string): Promise<unknown> =>
-      ipcRenderer.invoke(IpcChannels.collaboration.prompt, sessionId, workspaceId, query, clientMessageId),
-
-    onEvent: (callback: (event: import('../src/types/collaboration').CollaborationEvent) => void): (() => void) => {
-      const handler = (_event: Electron.IpcRendererEvent, data: import('../src/types/collaboration').CollaborationEvent) => {
-        callback(data);
-      };
-      ipcRenderer.on(IpcChannels.collaboration.event, handler);
-      return () => {
-        ipcRenderer.removeListener(IpcChannels.collaboration.event, handler);
-      };
-    },
-  },
+  collaboration: collaborationBridge,
 
   subagent: subagentBridge,
   skills: skillsBridge,
