+++ apps/desktop/src/stores/subagent-constants.ts
@@ -1,0 +1,2 @@
+/** Terminal statuses that count as "done" for subagent entries. */
+export const TERMINAL_STATUSES = new Set(['completed', 'failed', 'aborted', 'timed_out']);
